[APPINT-1205] Fix cron jobs
The previous implementation was randomizing connector start time at
every initialization by clearing the cron and recreating it.
This causes issues when debugging and trying to pause a job as the cron
config is reset anytime Rails is initialized (rake, console, ...)

It's enought to generate a random time when generating the connector.