[APPINT-1146] Fix database encoding configuration
Sometimes the database is created with an character set different that
the one specified in `config/database.yml`.
This causes some issue when the database is created in non utf8 charset
(eg: latin1).

This commit adds an initializer that warn the user in case of mismatch
between the configured and actual encoding.
It also provides a migration generator to facilitate charset migration.